Fix performance issue in CalendarTemplate uses. Fixed some bugs in Calendar with leap years and size Added basic unit test for Calendar and CalendarTemplate

+++ src/10_db/RankUpdateQuery.hpp
@@ -69,9 +69,9 @@ namespace synthese
 			RankUpdateQuery(
 				const std::string& rankColumn,
 				int delta,
-				int rank,
+				size_t rank,
 				bool orEqual = true
-			) {
+			){
 				_updateQuery.reset(new UpdateQuery<Table>());
 
 				_updateQuery->addUpdateField(rankColumn, RawSQL(rankColumn + " + " + boost::lexical_cast<string>(delta)));
